Split cameras into separate files.
